Fix some migration bugs

+++ fuel/app/migrations/049_delete_top_parent_division_id_from_divisions.php
@@ -6,19 +6,35 @@ class Delete_top_parent_division_id_from_divisions
 {
 	public function up()
 	{
-		\DBUtil::drop_index('divisions', 'idx_divisions_top_parent_division_id');
-		\DBUtil::drop_fields('divisions', array(
-			'belongs_division_id',
-			'top_parent_division_id'
-		));
+		if ( \DBUtil::field_exists('divisions', array('top_parent_division_id')))
+		{
+			\DBUtil::drop_index('divisions', 'idx_divisions_top_parent_division_id');
+			\DBUtil::drop_fields('divisions', array(
+				'top_parent_division_id'
+			));
+		}
+		if ( \DBUtil::field_exists('divisions', array('belongs_division_id')))
+		{
+			\DBUtil::drop_fields('divisions', array(
+				'belongs_division_id',
+			));
+		}
 	}
 
 	public function down()
 	{
-		\DBUtil::add_fields('divisions', array(
-			'belongs_division_id' => array('constraint' => 50,  'null' => true, 'type' => 'varchar'),
-			'top_parent_division_id' => array('constraint' => 11,  'null' => true, 'type' => 'int'),
-		));
-		\DBUtil::create_index('divisions', 'top_parent_division_id', 'idx_divisions_top_parent_division_id');
+		if ( ! \DBUtil::field_exists('divisions', array('belongs_division_id')))
+		{
+			\DBUtil::add_fields('divisions', array(
+				'belongs_division_id' => array('constraint' => 11,  'null' => true, 'type' => 'int'),
+			));
+		}
+		if ( ! \DBUtil::field_exists('divisions', array('top_parent_division_id')))
+		{
+			\DBUtil::add_fields('divisions', array(
+				'top_parent_division_id' => array('constraint' => 11,  'null' => true, 'type' => 'int'),
+			));
+			\DBUtil::create_index('divisions', 'top_parent_division_id', 'idx_divisions_top_parent_division_id');
+		}
 	}
 }
